+++ printer-driver-api/src/main/java/com/aevi/print/driver/common/PrinterDriverFactory.java
@@ -13,7 +13,7 @@
  */
 package com.aevi.print.driver.common;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.aevi.print.model.BasePrinterInfo;